+++ lib/react_native_util/converter.rb
@@ -4,33 +4,12 @@ require 'json'
 require 'rubygems'
 require 'tmpdir'
 require 'tty/platform'
-require 'xcodeproj'
-require_relative 'util'
+require_relative 'project'
 
 module ReactNativeUtil
   # Class to perform conversion operations.
   class Converter
     include Util
-
-    # [Array<String>] Xcode projects from react-native that may be in the Libraries group
-    DEFAULT_DEPENDENCIES = %w[
-      ART
-      React
-      RCTActionSheet
-      RCTAnimation
-      RCTBlob
-      RCTCameraRoll
-      RCTGeolocation
-      RCTImage
-      RCTLinking
-      RCTNetwork
-      RCTPushNotification
-      RCTSettings
-      RCTTest
-      RCTText
-      RCTVibration
-      RCTWebSocket
-    ]
 
     # [String] Path to the Podfile template
     PODFILE_TEMPLATE_PATH = File.expand_path '../assets/templates/Podfile.erb', __dir__
@@ -41,10 +20,11 @@ module ReactNativeUtil
     # [String] Full path to Xcode project
     attr_reader :xcodeproj_path
 
-    # [Xcodeproj::Project] Contents of the project at xcodproj_path
-    attr_reader :xcodeproj
+    # [Project] Contents of the project at xcodeproj_path
+    attr_reader :project
 
     attr_reader :options
+    attr_reader :react_project
     attr_reader :react_podspec
 
     def initialize(repo_update: nil)
@@ -91,7 +71,7 @@ module ReactNativeUtil
       load_xcodeproj!
       log "Found Xcode project at #{xcodeproj_path}"
 
-      if libraries_group.nil?
+      if project.libraries_group.nil?
         log "Libraries group not found in #{xcodeproj_path}. No conversion necessary."
         exit 0
       end
@@ -103,35 +83,37 @@ module ReactNativeUtil
         exit 1
       end
 
+      project.validate_app_target!
+
       load_react_podspec!
 
       # 2. Detect native dependencies in Libraries group.
       log 'Dependencies:'
-      dependencies.each { |d| log " #{d}" }
+      project.dependencies.each { |d| log " #{d}" }
 
       # Save for after Libraries removed.
-      deps_to_add = dependencies
+      deps_to_add = project.dependencies
 
       # 3. Run react-native unlink for each one.
       log 'Unlinking dependencies'
-      dependencies.each do |dep|
+      project.dependencies.each do |dep|
         run_command_with_spinner! 'react-native', 'unlink', dep, log: File.join(Dir.tmpdir, "react-native-unlink-#{dep}.log")
       end
 
       # reload after react-native unlink
       load_xcodeproj!
+      load_react_project!
 
       # 4a. Add Start Packager script
-      validate_app_target!
-      add_packager_script
+      project.add_packager_script_from react_project
 
       # Make a note of pod subspecs to replace Libraries group
       load_subspecs_from_libraries
 
       # 4b. Remove Libraries group from Xcode project.
-      remove_libraries_group_from_project!
+      project.remove_libraries_group
 
-      xcodeproj.save
+      project.save
 
       # 5. Generate boilerplate Podfile.
       generate_podfile!
@@ -223,8 +205,9 @@ module ReactNativeUtil
     # Load the project at @xcodeproj_path into @xcodeproj
     # @raise ConversionError on failure
     def load_xcodeproj!
-      @xcodeproj = nil # in case of exception on reopen
-      @xcodeproj = Xcodeproj::Project.open xcodeproj_path
+      @project = nil # in case of exception on reopen
+      @project = Project.open xcodeproj_path
+      @project.app_name = app_name
     rescue Errno::ENOENT
       raise ConversionError, "Failed to open #{xcodeproj_path}. File not found."
     rescue Xcodeproj::PlainInformative => e
@@ -242,72 +225,9 @@ module ReactNativeUtil
       # rubocop: enable Security/Eval
     end
 
-    # A representation of the Libraries group (if any) from the Xcode project.
-    # @return the Libraries group
-    def libraries_group
-      xcodeproj['Libraries']
-    end
-
-    # Remove the Libraries group from the xcodeproj in memory.
-    # Resets @libraries_group to nil as well.
-    def remove_libraries_group_from_project!
-      # Remove links against these static libraries
-      xcodeproj.targets.reject { |t| t.name =~ /-tvOS/ }.each do |t|
-        remove_libraries_from_target t
-      end
-
-      log 'Removing Libraries group'
-      libraries_group.remove_from_project
-    end
-
-    def remove_libraries_from_target(target)
-      log "Removing Libraries from #{target.name}"
-      to_remove = target.frameworks_build_phase.files.select do |file|
-        path = file.file_ref.pretty_print
-        next false unless /^lib(.+)\.a$/.match?(path)
-
-        static_libs.include?(path)
-      end
-
-      to_remove.each { |f| target.frameworks_build_phase.remove_build_file f }
-    end
-
-    # A list of external dependencies from NPM requiring react-native link.
-    # @return [Array<String>] a list of NPM package names
-    def dependencies
-      return [] if libraries_group.nil?
-
-      dependency_paths.map do |path|
-        # Find the root above the ios/*.xcodeproj under node_modules
-        root = File.expand_path '../..', path
-        File.basename root
-      end
-    end
-
-    # Paths to Xcode projects in the Libraries group from external deps.
-    # @return [Array<String>] a list of absolute paths to Xcode projects
-    def dependency_paths
-      return [] if libraries_group.nil?
-
-      paths = libraries_group.children.reject { |c| DEFAULT_DEPENDENCIES.include?(c.name.sub(/\.xcodeproj$/, '')) }.map(&:path)
-      paths.map { |p| File.expand_path p, File.join(Dir.pwd, 'ios') }
-    end
-
-    def library_roots
-      libraries_group.children.map do |library|
-        File.basename(library.path).sub(/\.xcodeproj$/, '')
-      end
-    end
-
-    # All static libraries from the Libraries group
-    # @return [Array<String>] an array of filenames
-    def static_libs
-      library_roots.map { |root| "lib#{root}.a" }
-    end
-
     def load_subspecs_from_libraries
-      roots = library_roots - %w[React]
-      @subspecs_from_libraries = roots.select { |r| DEFAULT_DEPENDENCIES.include?(r) }.map do |root|
+      roots = project.library_roots - %w[React]
+      @subspecs_from_libraries = roots.select { |r| Project::DEFAULT_DEPENDENCIES.include?(r) }.map do |root|
         case root
         when 'RCTLinking'
           'RCTLinkingIOS'
@@ -324,18 +244,11 @@ module ReactNativeUtil
     end
 
     def app_target
-      xcodeproj.targets.find { |t| t.name == app_name }
+      project.app_target
     end
 
     def test_target
-      xcodeproj.targets.select(&:test_target_type?).reject { |t| t.name =~ /tvOS/ }.first
-    end
-
-    # Validate an assumption about the project. TODO: Provide override option.
-    # @raise ConversionError if an application target is not found with the same name as the project.
-    def validate_app_target!
-      raise ConversionError, "Unable to find target #{app_name} in #{xcodeproj_path}." if app_target.nil?
-      raise ConversionError, "Target #{app_name} is not an application target." unless app_target.product_type == 'com.apple.product-type.application'
+      project.test_target
     end
 
     def podfile_path
@@ -369,55 +282,11 @@ module ReactNativeUtil
       raise ConversionError, 'Uncommitted changes in repo. Please commit or stash before continuing.'
     end
 
-    # Adds the Start Packager script from the React.xcodeproj under node_modules
-    # to the main application target before deleting React.xcodeproj from the
-    # Libraries group. Adjusts paths in the script to account for the different
-    # project location. If React.xcodeproj cannot be opened, or if the relevant
-    # build phase is not found, a warning is logged, and this step is skipped.
-    #
-    # TODO: The build phase is added after all other build phases. Ideally it
-    # can be moved to the beginning. The packager is independent of the Xcode
-    # build process. It may be started at any time. Starting it early is an
-    # optimization that allows it to load while the build is in progress.
-    # Currently it's possible to simply drag the build phase in Xcode to a
-    # higher position after running the react_pod command.
-    def add_packager_script
-      old_packager_phase = packager_phase_from_react_project
-      unless old_packager_phase
-        log 'Could not find packager build phase in React.xcodeproj. Skipping.'.yellow
-        return
-      end
-
-      # location of project is different relative to packager script
-      script = old_packager_phase.shell_script.gsub(%r{../scripts}, '../node_modules/react-native/scripts')
-
-      phase = app_target.new_shell_script_build_phase old_packager_phase.name
-      phase.shell_script = script
-    end
-
-    # Returns a Project object with the contents of the React.xcodeproj project
-    # from node_modules.
-    # @return [Xcodeproj::Project] a Project object with the contents of React.xcodeproj from node_modules
+    # Load the contents of the React.xcodeproj project from node_modules.
     # @raise Xcodeproj::PlainInformative in case of most failures
-    def react_project!
-      return @react_project if @react_project
-
-      path = libraries_group.children.find { |c| c.path =~ /React.xcodeproj/ }.real_path
+    def load_react_project!
+      path = project.libraries_group.children.find { |c| c.path =~ /React.xcodeproj/ }.real_path
       @react_project = Xcodeproj::Project.open path
-    end
-
-    # Returns the original Start Packager build phase from the React.xcodeproj
-    # under node_modules. This contains the original script.
-    # @return the packager build phase if found
-    # @return nil if not found or React.xcodeproj cannot be opened
-    def packager_phase_from_react_project
-      react_project!.targets.first.build_phases.find { |p| p.name =~ /packager/i }
-    rescue Errno::ENOENT
-      log 'Could not open React.xcodeproj. File not found.'
-      nil
-    rescue Xcodeproj::PlainInformative => e
-      log "Could not open React.xcodeproj. #{e.message}"
-      nil
     end
   end
 end
